lesson4 feat: add basic routing to home, chats, profile pages
