+++ src/Profile.js
@@ -1,0 +1,5 @@
+function Profile(props) {
+  return <p>Profile page</p>;
+}
+
+export default Profile;
